% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biodatascience}
\alias{biodatascience}
\title{Collected data of interactive quizzes}
\format{A data frame with 45721 rows and 11 variables:
\describe{
\item{date}{Time of entry}
\item{tutorial}{Quiz currently completed}
\item{version}{Version of the quiz}
\item{user_name}{Generic (anonymized) name of the student}
\item{label}{Label of the question in the quiz}
\item{correct}{For a multiple choice questions, whether answer is correct
(\code{TRUE}) or not (\code{FALSE}). For an open question, whether the
student submitted its final solution (\code{TRUE}), no mather if that
solution is correct or not.}
\item{event}{Events triggering the entry (learnr UI-related).}
\item{data}{The answers proposed by the student (JSON string).}
\item{tuto_label}{Combination of tutorial and label variables.}
\item{data_conv}{Answer extracted of 'data'.}
\item{data_qu}{Question extracted of 'data'.}
}}
\source{
\url{http://biodatascience-course.sciviews.org}
}
\usage{
biodatascience
}
\description{
Collected data of interactive quizzes used in a biological data science course.
}
\examples{
data(biodatascience)
class(biodatascience)
head(biodatascience)

}
\keyword{datasets}
